+++ app/Http/Controllers/CityController.php
@@ -83,6 +83,11 @@ class CityController extends Controller
     public function destroy(string $id)
     {
         $row = City::findorFail($id);
+
+        if($row->applications()->count() > 0){
+            return Response(['result'=>'error','message'=>__('Data for this city already exists')]);
+        }
+
         $row->delete();
 
         return Response(['result'=>'success','message'=>__('Deleted Successfully')]);
@@ -108,6 +113,7 @@ class CityController extends Controller
 
         //-- CREATE LARAVEL PAGINATION
         $paginate =  City::search($conditions)
+                ->latest()
                 ->orderBy($columnName, $columnSortOrder)
                 ->paginate($limit, ["*"], 'page', $page);
